fix: Build output directory structure (#3)
The extension is now compiled to the root of the `out` directory.

Before this PR the `out` directory contained `src` and `test`
subdirectories. The `extension.ts` was compiled to
`out/src/extension.js`, which was not compatible with `main` definition
in the `package.json`.

## Changes

Excluded tests from `tsconfig.json`. Now there is no need for a separate
`tsconfig.compile.json`.